change to avoid compilation warnings from gcc 4.1.2: virtual destructors, assert
git-svn-id: https://svn.resiprocate.org/rep/resiprocate@8128 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ branches/b-teltel_penguin_server-20050706/resiprocate/MsgHeaderScanner.cxx
@@ -935,7 +935,7 @@ enum { chunkTermSentinelChar = '\0' };
             localTextPropBitMask |= charInfo->textPropBitMask;
           determineTransitionFromCharCategory:
             TransitionInfo *transitionInfo =
-                &(localStateMachine[localState][charCategory]);
+	       &(localStateMachine[localState][(int)charCategory]);
             TransitionAction transitionAction = transitionInfo->action;
 #if defined(RESIP_MSG_HEADER_SCANNER_DEBUG)  
             printStateTransition(localState, *charPtr, transitionAction);
